pyproject.toml
[tool.poetry]
name = "sql2spl"
version = "0.1.0"
description = "translate sql to spl"
authors = ["Jiangtao Peng <pengjtnina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
antlr4-python3-runtime = "^4.13.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
antlr4-tools = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
